Move components to svelte folder
