+++ src/svelte/ButtonLink.svelte
@@ -1,10 +1,11 @@
 <script>
-  export let href;
-  import Link from './independent/Link.svelte';
+  export let href = '';
+  export let disabled = false;
+  import Link from './Link.svelte';
 </script>
 
-<span>
-  <Link {href} invert={false}>
+<span class:disabled>
+  <Link href={!disabled ? href : ''} invert={false}>
     <slot />
   </Link>
 </span>
@@ -12,7 +13,7 @@
 <style>
   span {
     display: inline-flex;
-    border-radius: 4px;
+    border-radius: 0.25em;
     text-transform: uppercase;
     color: var(--fg-color);
     background-color: var(--bg-color);
@@ -23,6 +24,19 @@
     box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.2), 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12);
     transform: scale(1.05);
   }
+
+  span.disabled {
+    pointer-events: none;
+    cursor: not-allowed;
+    opacity: 0.4;
+    text-transform: none;
+  }
+  span.disabled:hover,
+  span.disabled:active {
+    box-shadow: none;
+    transform: none;
+  }
+
   span :global(a) {
     padding: 0.5em 1em;
   }
