+++ src/svelte/Searchbar.svelte
@@ -1,7 +1,7 @@
 <script>
   export let query = '';
   export let filters = true;
-  import Icon from './independent/Icon.svelte';
+  import Icon from './Icon.svelte';
   import { createEventDispatcher } from 'svelte';
   const dispatch = createEventDispatcher();
 </script>
